[refactor] cn-user-api lib.rs / contract.rs をドメイン別に分割 (WP-H4)
lib.rs(1,331 行)を分割(公開 re-export 維持で呼び出し側変更ゼロ):
- config / state(DI)/ rate_limit / routes / errors +
  handlers 6 ドメイン(auth / consents / reports / indexing /
  trust_relation / bootstrap)
- apply_rate_limit の重複 2 分岐(key extractor の型違いのみ)を
  generic ヘルパ 1 本に統合
- parse_u64_env / parse_u32_env を cn-core::env へ昇格
  (parse_bool_env / parse_csv_env の隣)

tests/contract.rs(1,207 行)を分割:
- tests/support/mod.rs に共有ヘルパ(TestServer / authenticate 系 /
  consents 受諾 / redis 検査 / 統合テスト gate)を集約
- contract_auth.rs(auth + admission)/ contract_bootstrap.rs
  (bootstrap + rendezvous)へ分割
- 兄弟テスト 4 本(reports / index_query / indexing_requests /
  trust_relation)の重複ヘルパを support 参照へ置換

挙動不変(HTTP contract 不変)。oversized baseline から 2 件解消。
