+++ crates/cn-user-api/src/handlers/mod.rs
@@ -1,0 +1,8 @@
+//! ドメイン別ハンドラ(routes.rs から参照される)。
+
+pub(crate) mod auth;
+pub(crate) mod bootstrap;
+pub(crate) mod consents;
+pub(crate) mod indexing;
+pub(crate) mod reports;
+pub(crate) mod trust_relation;
