+++ crates/cn-user-api/tests/reports.rs
@@ -11,8 +11,8 @@ use kukuri_cn_core::{JwtConfig, TestDatabase};
 use kukuri_cn_user_api::{UserApiConfig, app_router, build_state};
 use reqwest::{Client, StatusCode};
 
-const DEFAULT_ADMIN_DATABASE_URL: &str = "postgres://cn:cn_password@127.0.0.1:15432/cn";
-const DEFAULT_RENDEZVOUS_REDIS_URL: &str = "redis://127.0.0.1:16379/";
+mod support;
+use support::{integration_test_admin_database_url, integration_test_rendezvous_redis_url};
 
 /// report_endpoint capability を有効化した operator config。
 const REPORT_ENABLED_YAML: &str = r#"server:
@@ -96,29 +96,6 @@ impl TestServer {
         self.task.abort();
         self.database.cleanup().await
     }
-}
-
-fn integration_test_admin_database_url() -> Option<String> {
-    let enabled = std::env::var("KUKURI_CN_RUN_INTEGRATION_TESTS")
-        .ok()
-        .map(|value| matches!(value.trim(), "1" | "true" | "TRUE" | "yes" | "YES"))
-        .unwrap_or(false);
-    if !enabled {
-        return None;
-    }
-    Some(
-        std::env::var("COMMUNITY_NODE_DATABASE_URL")
-            .ok()
-            .filter(|value| !value.trim().is_empty())
-            .unwrap_or_else(|| DEFAULT_ADMIN_DATABASE_URL.to_string()),
-    )
-}
-
-fn integration_test_rendezvous_redis_url() -> String {
-    std::env::var("COMMUNITY_NODE_RENDEZVOUS_REDIS_URL")
-        .ok()
-        .filter(|value| !value.trim().is_empty())
-        .unwrap_or_else(|| DEFAULT_RENDEZVOUS_REDIS_URL.to_string())
 }
 
 #[tokio::test]
